pyproject.toml
[project]
name = "boardgame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pygame-ce>=2.5.2",
]
